SS 4.0 - Add 'TestOnly' implementation to test objects, update extraDataObjects to extra_dataobjects static variable
