+++ tests/QueuedJobsTest/TestQJService.php
@@ -1,11 +1,12 @@
 <?php
 
-namespace SilverStripe\QueuedJobs\Tests;
+namespace SilverStripe\QueuedJobs\Tests\QueuedJobsTest;
 
 use SilverStripe\QueuedJobs\Services\QueuedJobService;
+use SilverStripe\Dev\TestOnly;
 
 // stub class to be able to call init from an external context
-class TestQJService extends QueuedJobService
+class TestQJService extends QueuedJobService implements TestOnly
 {
     /**
      * Not inherited from QueuedJobService unfortunately...
